Move code to shopping-list module
