+++ client/src/app/model/reducer.js
@@ -1,8 +1,11 @@
 import { combineReducers } from 'redux';
 
-import items, { uiStatus } from 'modules/shopping-list/model/reducer';
+import items, {
+  uiStatus,
+  shoppingLists
+} from 'modules/shopping-list/model/reducer';
 import currentUser from 'modules/authorization/model/reducer';
-import { shoppingLists, cohortsList } from 'modules/dashboard/model/reducer';
+import { cohortsList } from 'modules/dashboard/model/reducer';
 
 const rootReducer = combineReducers({
   cohortsList,
