+++ client/src/app/styles/index.scss
@@ -34,4 +34,4 @@
 @import '../components/Toolbar/Toolbar.scss';
 @import '../../modules/dashboard/Dashboard.scss';
 @import '../../modules/dashboard/CardItem.scss';
-@import '../../app/components/Toolbar/components/CreationForm.scss';
+@import '../../modules/shopping-list/components/CreationForm.scss';
