+++ client/src/modules/shopping-list/components/CreationForm.scss
